Moved classes to new package locations.
