+++ app/src/main/java/com/android/shawnclisby/thetodolist/data/TaskViewModelFactory.kt
@@ -1,9 +1,8 @@
-package com.android.shawnclisby.thetodolist
+package com.android.shawnclisby.thetodolist.data
 
 import android.app.Application
 import androidx.lifecycle.ViewModel
 import androidx.lifecycle.ViewModelProvider
-import com.android.shawnclisby.thetodolist.data.TaskViewModel
 
 class TaskViewModelFactory(private val application: Application) : ViewModelProvider.Factory {
     override fun <T : ViewModel?> create(modelClass: Class<T>): T {
